+++ cashocs/_optimization/shape_optimization/mesh_constraint_manager.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020-2023 Sebastian Blauth
+# Copyright (C) 2020-2024 Sebastian Blauth
 #
 # This file is part of cashocs.
 #
@@ -98,7 +98,7 @@ class ConstraintManager:
                 )
 
             elif self.mesh.geometry().dim() == 3:
-                angle_constraint = mesh_constraints.DihedralAngleConstraint(
+                angle_constraint = mesh_constraints.SolidAngleConstraint(
                     self.mesh, self.config, deformation_space
                 )
             else: